+++ compat-test/express-zod-api.ts
@@ -1,6 +1,7 @@
 import { createConfig, createServer } from "express-zod-api";
 import { Server } from "socket.io";
-import { attachSockets, createSimpleConfig, Integration } from "zod-sockets";
+import { attachSockets, createSimpleConfig } from "zod-sockets";
+import { Integration } from "zod-sockets/integration";
 import { givePort } from "../tools/ports.ts";
 
 const serverConfig = createConfig({
@@ -8,18 +9,16 @@ const serverConfig = createConfig({
   cors: false,
 });
 
-const { servers, logger } = await createServer(serverConfig, {});
+const { servers, logger } = createServer(serverConfig, {});
 
 const socketsConfig = createSimpleConfig();
 
 logger.info(
   "Integration sample",
-  (
-    await Integration.create({
-      config: socketsConfig,
-      actions: [],
-    })
-  ).print(),
+  new Integration({
+    config: socketsConfig,
+    actions: [],
+  }).print(),
 );
 
 const io = new Server();